Improve time/timestamp presentation
Helpers/functions that convert timestamp to human readable form are
prefixed with 'timestamp-' while others are prefixed with 'time-'.
In the UI, the time when an object was last modified is shown in
`xxx ago` format, while the exact date/time when it was modified is
shown on mouse hover.